Automaton

Using these types and methods, you can build a Finite State Automaton.

Liblet.Automaton — Type

This type represents a (nondeterministic) finite automaton, defined as a tuple $A=(N,T,transitions,q0,F)$ Automaton(N::Iterable, T::Iterable, transitions::Iterable, q0::AbstractString, F::Iterable) Builds an Automaton obtained from the given components.

Liblet.Automaton — Type

Automaton(transitions::AbstractString, F::Union{Nothing,Set} = nothing, q0::Union{Nothing, AbstractString} = nothing)::Automaton

Builds an Automaton obtained from the given transitions.

Liblet.Automaton — Method

Automaton(G::Grammar)::Automaton

Builds the Automaton corresponding to the given regular Grammar.

Liblet.δ — Method

δ(a::Automaton, X, x)

The transition function.

This function returns the set of states reachable from the given state and input symbol.