Derivation

You can build derivations from a given Grammar.

Liblet.Derivation — Type

A derivation.

Liblet.Derivation — Method

Derivation(G::Grammar)::Derivation

Builds a Derivation from a given Grammar.

Liblet.next — Method

next(d::Derivation, prod::Int, pos::Int)::Derivation

Applies the specified production to the given position in the sentential form.

Liblet.next — Method

next(d::Derivation, prod::Production, pos::Int)::Derivation

Applies the specified Production to the given position in the sentential form.

Liblet.next — Method

next(d::Derivation, prod::Int, pos::Int)::Derivation

Applies the specified productions to the given position in the sentential form, one by one

Liblet.leftmost — Method

leftmost(d::Derivation, prod::Int)::Derivation

Performs a leftmost derivation step. Applies the specified production to the current leftmost nonterminal in the sentential form.

Liblet.leftmost — Method

leftmost(d::Derivation, prod::Production)::Derivation

Performs a leftmost derivation step. Applies the specified Production to the current leftmost nonterminal in the sentential form.

Liblet.leftmost — Method

leftmost(d::Derivation, prod::AbstractArray{Production})::Derivation

Performs a leftmost derivation step. Applies the specified productions to the current leftmost nonterminal in the sentential form, one by one.

Liblet.leftmost — Method

leftmost(d::Derivation, prod::AbstractArray{Int})::Derivation

Performs a leftmost derivation step. Applies the specified productions to the current leftmost nonterminal in the sentential form, one by one.

Liblet.rightmost — Method

rightmost(d::Derivation, prod::Int)::Derivation

Performs a rightmost derivation step. Applies the specified production(s) to the current rightmost nonterminal in the sentential form.

Liblet.rightmost — Method

rightmost(d::Derivation, prod::Production)::Derivation

Performs a rightmost derivation step. Applies the specified Production to the current rightmost nonterminal in the sentential form.

Liblet.rightmost — Method

rightmost(d::Derivation, prod::Int)::Derivation

Performs a rightmost derivation step. Applies the specified productions to the current rightmost nonterminal in the sentential form, one by one.

Liblet.rightmost — Method

rightmost(d::Derivation, prod::AbstractArray{Production})::Derivation

Performs a rightmost derivation step. Applies the specified productions to the current rightmost nonterminal in the sentential form, one by one.

Liblet.possiblesteps — Method

possiblesteps(d::Derivation; prod::Union{Int, Nothing} = nothing, pos::Union{Int, Nothing} = nothing)

Returns all the possible steps that can be performed given the grammar and current sentential form.

Determines all the position of the sentential form that correspond to the left-hand side of one of the production in the grammar, returning the position and production number. If a production is specified, it yields only the pairs referring to it; similarly, if a position is specified, it yields only the pairs referring to it.

Liblet.sententialform — Method

sententialform(d::Derivation)

Returns the sentential form of the Derivation.

Liblet.steps — Method

steps(d::Derivation)

Returns the steps performed by the Derivation