Grammar

Liblet.Grammar — Type

A grammar, represented as a tuple $G=(N,T,P,S)$

Liblet.Grammar — Method

Grammar(N::Set, T::Set, P::Array, S::AbstractString)::Grammar

Builds a Grammar obtained from the given string of productions.

Liblet.Grammar — Type

Grammar(prods::AbstractString, iscontextfree = true)::Grammar

Builds a Grammar obtained from the given string of productions.

Liblet.alternatives — Method

alternatives(g::Grammar, N::Array)::Array

Returns all the right-hand sides alternatives matching the given nonterminal.

Liblet.restrict — Method

restrict(g::Grammar, symbols::Set)

Returns a Grammar using only the given symbols.