Productions

Productions are the most basic thing in languages, translators and compilers.

Introduction

Liblet.AbstractProduction — Type

Supertype for productions. Production and other types are subtypes of this.

Production

Liblet.Production — Type

A grammar production, consisting of a left side and a right side

Liblet.parseproduction — Method

parseproduction(input::AbstractString, iscontextfree::Bool = true)::Array{Production}

Returns an Array of Production obtained from the given string.

Liblet.suchthat — Method

suchthat(;left::AbstractString = nothing, right::AbstractString = nothing, rightlen::Int = nothing, right_is_suffix_of::AbstractString = nothing)::Array

A predicate (that is a one-argument function that retuns true or false) that is true weather the production given as argument satisfies all the predicates given by the named arguments.

Liblet.astype0 — Method

astype0(p::Production)::Production

Returns a new Production that is type 0

Item

For the purpose of presenting the Knuth Automaton in the context of LR(0) parsing, Item is an extended Production that includes a dot.

Liblet.Item — Type

A dotted production, also known as an Item.

Liblet.parseitem — Method

parseitem(input::AbstractString, iscontextfree::Bool = true)::Array{Item}

Returns an Array of Item obtained from the given string.

Liblet.astype0 — Method

astype0(p::Item)::Item

Returns a new Item that is type 0

Liblet.afterdotsymbol — Method

afterdotsymbol(item::Item)::Union{AbstractString,Nothing}

Returns the symbol after the dot.

Liblet.advance — Method

advance(item::Item, x::AbstractString)::Item

Returns a new Item obtained advancing the dot past the given symbol.